// src/app/products/interfaces/productsByCategory.interface.ts
// Generated by https://quicktype.io

export interface ProductsByCategory {
  products: Product[];
  total: number;
  skip: number;
  limit: number;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}