// src/app/shared/interfaces/carts.interfaces.ts
// Generated by https://quicktype.io

export interface CartItem {
  productId: number;
  title: string;
  price: number;
  quantity: number;
  total: number;
  thumbnail: string;
  description: string;
}

export interface ShoppingCart {
  items: CartItem[];
  totalItems: number;
  totalCost: number;
}